Move unchecked header import out of Verifier struct
